+++ src/Query/Prefix.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Elastica\Query;
 
+use Elastica\Exception\InvalidException;
+
 /**
  * Prefix query.
  *
@@ -11,6 +13,21 @@ namespace Elastica\Query;
  */
 class Prefix extends AbstractQuery
 {
+    /**
+     * Rewrite methods: @see https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-multi-term-rewrite.html.
+     */
+    public const REWRITE_CONSTANT_SCORE = 'constant_score';
+    public const REWRITE_CONSTANT_SCORE_BOOLEAN = 'constant_score_boolean';
+    public const REWRITE_SCORING_BOOLEAN = 'scoring_boolean';
+    public const REWRITE_TOP_TERMS_BLENDED_FREQS_N = 'top_terms_blended_freqs_N';
+    public const REWRITE_TOP_TERMS_BOOST_N = 'top_terms_boost_N';
+    public const REWRITE_TOP_TERMS_N = 'top_terms_N';
+
+    /**
+     * @var string|null
+     */
+    private $field;
+
     /**
      * @param array $prefix OPTIONAL Calls setRawPrefix with the given $prefix array
      */
@@ -29,6 +46,8 @@ class Prefix extends AbstractQuery
      */
     public function setRawPrefix(array $prefix): self
     {
+        $this->field = \array_key_first($prefix);
+
         return $this->setParams($prefix);
     }
 
@@ -44,5 +63,23 @@ class Prefix extends AbstractQuery
     public function setPrefix(string $key, $value, float $boost = 1.0): self
     {
         return $this->setRawPrefix([$key => ['value' => $value, 'boost' => $boost]]);
+    }
+
+    /**
+     * Set the method used to rewrite the query.
+     * Use one of the Prefix::REWRITE_* constants, or provide your own.
+     *
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-multi-term-rewrite.html
+     */
+    public function setRewrite(string $rewriteMode): self
+    {
+        if (null === $this->field) {
+            throw new InvalidException('No field has been set');
+        }
+
+        $data = $this->getParam($this->field);
+        $this->setParam($this->field, \array_merge($data, ['rewrite' => $rewriteMode]));
+
+        return $this;
     }
 }